refactor: ファイルの管理位置を修正
